Display a selector field and populate with the following values. - Simple bar - value-> type: bar - Horizontal bar - value -> type: bar and direction: horizontal - Simple scatter - value: series type: scatter. - Pie - value: series type:pie
Also a input for X axis, and a input for Y axis and populate them with the selected field value. Display a small icon button X that helps to clear the value.

When a data field is selected, starting filling the X and Y input wwith the selected field.
Build a drag and drop UI so that users can swap X and Y fields.

To render visualization, determine the chart type selected and pass the value to visuzliation configuration. Display the name of the field for X and Y axis as labels.